Change over SST::RNG::SSTRandom to SST::RNG::Random (#644)
* Change over SST::RNG::SSTRandom to SST::RNG::Random as requested in Issue #625

* Reformatted per the existing SST clang-format file provided in the core repository.

* Removes rng.cc per @feldergast request.

* Include sstrng.h so we can accomodate bad coding practices
