+++ src/sst/core/rng/discrete.h
@@ -58,7 +58,7 @@ class SSTDiscreteDistribution : public SSTRandomDistribution {
         \param lambda The lambda of the exponential distribution
         \param baseDist The base random number generator to take the distribution from.
     */
-    SSTDiscreteDistribution(const double* probs, const uint32_t probsCount, SSTRandom* baseDist) :
+    SSTDiscreteDistribution(const double* probs, const uint32_t probsCount, SST::RNG::Random* baseDist) :
         probCount(probsCount) {
 
         probabilities = (double*) malloc(sizeof(double) * probsCount);
@@ -106,7 +106,7 @@ class SSTDiscreteDistribution : public SSTRandomDistribution {
         /**
             Sets the base random number generator for the distribution.
         */
-        SSTRandom* baseDistrib;
+        SST::RNG::Random* baseDistrib;
 
         /**
             Controls whether the base distribution should be deleted when this class is destructed.
