+++ src/sst/core/rng/sstrng.h
@@ -13,53 +13,14 @@
 #define SST_CORE_RNG_SSTRNG_H
 
 #include <stdint.h>
+#include "sst/core/rng/rng.h"
 
 namespace SST {
 namespace RNG {
 
-/**
-    \class SSTRandom sstrng.h "sst/core/rng/sstrng.h"
+using SSTRandom = SST::RNG::Random;
 
-    Implements the base class for random number generators for the SST core. This does not
-    implement an actual RNG itself only the base class which describes the methods each
-    class will implement.
-*/
-class SSTRandom {
+} // namespace RNG
+} // namespace SST
 
-    public:
-    /**
-        Generates the next random number in the range 0 to 1.
-    */
-    virtual double   nextUniform() = 0;
-
-    /**
-        Generates the next random number as an unsigned 32-bit integer.
-    */
-    virtual uint32_t generateNextUInt32() = 0;
-
-    /**
-        Generates the next random number as an unsigned 64-bit integer.
-    */
-    virtual uint64_t generateNextUInt64() = 0;
-
-    /**
-        Generates the next random number as a signed 64-bit integer.
-    */
-    virtual int64_t     generateNextInt64() = 0;
-
-    /**
-        Generates the next random number as a signed 32-bit integer
-    */
-        virtual int32_t  generateNextInt32() = 0;
-
-    /**
-        Destroys the random number generator
-    */
-    virtual ~SSTRandom() { }
-
-};
-
-} //namespace RNG
-} //namespace SST
-
-#endif //SST_CORE_RNG_SSTRNG_H
+#endif // SST_CORE_RNG_SSTRNG_H
